ContentProviderQueryable and FSDefaultProvider 0.13.0-beta1
0.13.0-beta1 of forsuredb changed the FSSelection interface
to return Object[] selectionArgs to be bound to an SQL query.
This change exposed a limitation of the Android sqlite
package that would not allow non-string arguments to be bound
to a SELECT query unless making a simple query for a single
value. Since the ContentProvider API only allows String type
selectionArgs, this workaround preserves both type and value
integrity through that limitation.